pyproject.toml
[tool.poetry]
name = "work1"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
aiogram = "^3.7.0"
aiosqlite = "^0.20.0"
pytoniq-core = "^0.1.36"
pytonconnect = "^0.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
